feature[ats_M5LLE]: Implement Attribute methods

+++ plugins/org.eclipse.osee.orcs.core/src/org/eclipse/osee/orcs/core/ds/DataProxy.java
@@ -30,7 +30,7 @@ public interface DataProxy {
 
    public Object[] getData() throws OseeCoreException;
 
-   public void persist(int storageId) throws OseeCoreException;
+   public void persist(long storageId) throws OseeCoreException;
 
    public void purge() throws OseeCoreException;
 }
