+++ plugins/org.eclipse.osee.orcs.db/src/org/eclipse/osee/orcs/db/internal/proxy/VarCharDataProxy.java
@@ -109,7 +109,7 @@ public class VarCharDataProxy extends AbstractDataProxy implements CharacterData
    }
 
    @Override
-   public void persist(int storageId) throws OseeCoreException {
+   public void persist(long storageId) throws OseeCoreException {
       getStorage().persist(storageId);
    }
